web/requirements.txt
cryptography==45.0.4
dotenv==0.9.9
fastapi==0.115.12
jinja2==3.1.6
matplotlib==3.10.3
nibabel==5.3.2
python-multipart==0.0.20
sqlalchemy==2.0.41
torch==2.6.0 --extra-index-url https://download.pytorch.org/whl/cu124
torchaudio==2.6.0 --extra-index-url https://download.pytorch.org/whl/cu124
torchinfo==1.8.0
torchview==0.2.7
torchvision==0.21.0 --extra-index-url https://download.pytorch.org/whl/cu124
uvicorn==0.34.3